// src/app/protected/customers/customers.interface.ts
export interface Customer {
    index: number;
    customerId: number;
    customerCMId: string;
    customerSocial: string;
    customerNameSurname: string;
    customerIdhome: string;
    customerAddress: string;
    customerTel: string;
    customerZipcode: string;
    customerEmail: string;
    customerIdcard: string;
    customerComment: string;
    customerDefualt: string;
    customerStatus: boolean;
    customerLastupdate: any;
    customerCreated: any;
    customerChannelId: string;
}